improved attention (#118)
* improved attention

* clean up lint

* format

+++ tests/nn/test_embeddings.py
@@ -13,7 +13,7 @@ import xax
 def test_embeddings_api(kind: xax.EmbeddingKind) -> None:
     key = jrandom.PRNGKey(0)
     xkey, pkey = jrandom.split(key)
-    x_tc = jrandom.normal(xkey, (5, 8), dtype=jnp.double)
+    x_tc = jrandom.normal(xkey, (5, 8), dtype=jnp.float32)
     times_t = jnp.arange(1, 6)
     emb = xax.get_positional_embeddings(kind, max_tsz=12, embed_dim=8, key=pkey)
     y1_tc = emb(x_tc, times_t=times_t)
@@ -26,7 +26,7 @@ def test_embeddings_api(kind: xax.EmbeddingKind) -> None:
 @pytest.mark.parametrize("offset", [0, 12])
 def test_rotary_embeddings_inference(offset: int) -> None:
     key = jrandom.PRNGKey(0)
-    x_tc = jrandom.normal(key, (5, 8), dtype=jnp.double)
+    x_tc = jrandom.normal(key, (5, 8), dtype=jnp.float32)
     emb = xax.get_positional_embeddings("rotary", max_tsz=8 + offset, embed_dim=8, learnable=False)
     y1_tc = emb(x_tc, offset=offset)
     y2_tc = xax.rotary_embeddings(x_tc, offset=offset)
